requirements.txt
discord.py[voice]~=2.6.4
pynacl~=1.5
python-dotenv~=1.2
yt-dlp==2025.10.22
bgutil-ytdlp-pot-provider==1.2.2
loguru~=0.7
pytest~=8.4
